server/requirements.txt
python-dotenv==1.0.0

openai==1.23.2
instructor==1.2.2
llama-index==0.10.30

fastapi==0.110.2
uvicorn==0.29.0
pydantic==2.7.0

pandas==2.2.2

deepgram-sdk==3.2.6